Using defaultBlockMap from Blocks index
