+++ lib/blocks.ts
@@ -1,15 +1,4 @@
-import { ReactNode } from 'react';
 import { ContentBlockAttribute } from '@/graphql/generated';
-import {
-	ClassicEditorBlock,
-	Heading,
-	BlockImage,
-	List,
-	Paragraph,
-	Quote,
-	Table,
-	UnsupportedBlock,
-} from '@/components/Blocks/index';
 
 /**
  * Map an array of ContentBlock attributes to an object that can used like props.
@@ -32,27 +21,6 @@ function mapAttributesToProps ( attributes: ContentBlockAttribute[] ): { [ key: 
 	}, {} );
 }
 
-type BlocksToComponentsProps = Record<string, ReactNode>;
-
-/**
- * Return an object with respected key-values based on Gutenberg Blocks
- * https://gogutenberg.com/blocks/
- */
-function mapBlockNamesToComponents ( override?: BlocksToComponentsProps ): Record<string, ReactNode> {
-	return {
-		'core/classic-editor': ClassicEditorBlock,
-		'core/heading': Heading,
-		'core/image': BlockImage,
-		'core/list': List,
-		'core/paragraph': Paragraph,
-		'core/quote': Quote,
-		'core/table': Table,
-		'unsupported': UnsupportedBlock,
-		...override,
-	}
-}
-
 export {
-	mapAttributesToProps,
-	mapBlockNamesToComponents,
+	mapAttributesToProps
 };
